Added some basic unit testing and fixed a pile of bugs for WordVectorizer.
Fixed MultiTokenCooccurrenceVectorizer to ensure that it returned a csr matrix.

Added a switch to RemoveEffectsTransformer that defaults to True to normalize any input that comes in.  In the future we may support this for counts but the present transform doesn't make any sense for non-normalized data.  The current fit step automatically normalizes in the PLSA.

Moved create_process_pipeline_stage and MultiTokenCooccurrenceVectorizer class to the utilities.py file in order to ensure that the class is defined before _MULTITOKEN_COOCCURRENCE_VECTORIZERS is instantiated.
