+++ textmap/vectorizers.py
@@ -5,7 +5,20 @@ from .transformers import (
     RemoveEffectsTransformer,
     MultiTokenExpressionTransformer,
 )
-from .tokenizers import NLTKTokenizer, BaseTokenizer, NLTKTweetTokenizer, SpacyTokenizer
+from .utilities import (
+    MultiTokenCooccurrenceVectorizer,
+    create_processing_pipeline_stage,
+    _INFO_WEIGHT_TRANSFORERS,
+    _REMOVE_EFFECT_TRANSFORMERS,
+)
+from .tokenizers import (
+    NLTKTokenizer,
+    BaseTokenizer,
+    NLTKTweetTokenizer,
+    SpacyTokenizer,
+    StanzaTokenizer,
+    SKLearnTokenizer,
+)
 from scipy.sparse import hstack
 from sklearn.preprocessing import normalize
 import pandas as pd
@@ -20,6 +33,8 @@ _TOKENIZERS = {
     "nltk": {"class": NLTKTokenizer, "kwds": {}},
     "tweet": {"class": NLTKTweetTokenizer, "kwds": {}},
     "spacy": {"class": SpacyTokenizer, "kwds": {}},
+    "stanza": {"class": StanzaTokenizer, "kwds": {}},
+    "sklearn": {"class": SKLearnTokenizer, "kwds": {}},
 }
 
 _CONTRACTORS = {
@@ -30,28 +45,17 @@ _CONTRACTORS = {
     "conservative": {"class": MultiTokenExpressionTransformer, "kwds": {}},
 }
 
-_COOCCURRENCE_VECTORIZERS = {
-    "symmetric": {"class": TokenCooccurrenceVectorizer, "kwds": {}},
-    "before": {
-        "class": TokenCooccurrenceVectorizer,
-        "kwds": {"window_orientation": "before"},
-    },
-    "after": {
-        "class": TokenCooccurrenceVectorizer,
-        "kwds": {"window_orientation": "after"},
-    },
-}
 
 _MULTITOKEN_COOCCURRENCE_VECTORIZERS = {
     "flat": {
-        "class": MultiTokenExpressionTransformer,
+        "class": MultiTokenCooccurrenceVectorizer,
         "kwds": {
             "vectorizer_list": ["before", "after"],
             "vectorizer_name_list": ["pre", "post"],
         },
     },
     "flat_1_5": {
-        "class": MultiTokenExpressionTransformer,
+        "class": MultiTokenCooccurrenceVectorizer,
         "kwds": {
             "vectorizer_list": ["before", "after", "before", "after"],
             "vectorizer_kwds_list": [
@@ -64,151 +68,6 @@ _MULTITOKEN_COOCCURRENCE_VECTORIZERS = {
         },
     },
 }
-
-_INFO_WEIGHT_TRANSFORERS = {
-    "default": {"class": InformationWeightTransformer, "kwds": {}},
-}
-
-_REMOVE_EFFECT_TRANSFORMERS = {
-    "defulat": {"class": RemoveEffectsTransformer, "kwds": {}},
-}
-
-
-def create_processing_pipeline_stage(class_to_create, class_dict, kwds, class_type):
-    if class_to_create is None:
-        return None
-    if class_to_create in class_dict:
-        _class = class_dict[class_to_create]["class"]
-        _kwds = class_dict[class_to_create]["kwds"]
-        if kwds is not None:
-            _kwds.update(kwds)
-        result = _class(**_kwds)
-    elif callable(class_to_create):
-        if kwds is not None:
-            result = class_to_create(**kwds)
-        else:
-            result = class_to_create()
-    else:
-        raise ValueError(
-            f"Unrecognized {class_type} {class_to_create}; should "
-            f"be one of {tuple(class_dict.keys())} or a {class_type} class."
-        )
-    return result
-
-
-class MultiTokenCooccurenceVectorizer(BaseEstimator, TransformerMixin):
-    """
-    Takes a sequence of token sequences, applies a set of TokenCooccurence views
-    of that data and returns the sparse concatination of all these views.
-
-    MultiTokenCooccurenceVectorizer(['before', 'after'],
-    """
-
-    def __init__(
-        self,
-        vectorizer_list,
-        vectorizer_kwds_list=None,
-        vectorizer_name_list=None,
-        info_weight_transformer="default",
-        info_weight_transformer_kwds=None,
-        remove_effects_transformer="default",
-        remove_effects_transformer_kwds=None,
-    ):
-        self.vectorizer_list = vectorizer_list
-        self.vectorizer_kwds_list = vectorizer_kwds_list
-        self.vectorizer_name_list = vectorizer_name_list
-        self.info_weight_transformer = info_weight_transformer
-        self.info_weight_transformer_kwds = info_weight_transformer_kwds
-        self.remove_effects_transformer = remove_effects_transformer
-        self.remove_effects_transformer_kwds = remove_effects_transformer_kwds
-
-    def fit(self, X):
-        """
-
-        Parameters
-        ----------
-        X: sequence of token sequences
-
-        Returns
-        -------
-
-        """
-        if self.vectorizer_name_list is None:
-            self.vectorizer_names_list_ = np.arange(len(self.vectorizer_list)).astype(
-                str
-            )
-        else:
-            self.vectorizer_names_list_ = self.vectorizer_name_list
-
-        if self.vectorizer_kwds_list is None:
-            self.vectorizer_kwds_list_ = [None] * len(self.vectorizer_list)
-        else:
-            self.vectorizer_kwds_list_ = self.vectorizer_kwds_list
-
-        self.info_weight_transformer_ = create_processing_pipeline_stage(
-            self.info_weight_transformer,
-            _INFO_WEIGHT_TRANSFORERS,
-            self.info_weight_transformer_kwds,
-            f"info weight transformer",
-        )
-
-        self.remove_effects_transformer_ = create_processing_pipeline_stage(
-            self.remove_effects_transformer,
-            _REMOVE_EFFECT_TRANSFORMERS,
-            self.remove_effects_transformer_kwds,
-            f"remove effects transformer",
-        )
-
-        # TODO: Check to make sure all matrices have the same number of rows and throw
-        # informative error.
-
-        self.column_dictionary_ = {}
-
-        for i, vectorizer in enumerate(self.vectorizer_list):
-            vectorizer_ = create_processing_pipeline_stage(
-                vectorizer,
-                _COOCCURRENCE_VECTORIZERS,
-                self.vectorizer_kwds_list_[i],
-                f"vectorizer {self.vectorizer_names_list_[i]}",
-            )
-            token_cooccurence = vectorizer_.fit_transform(X)
-            if self.info_weight_transformer_ is not None:
-                token_cooccurence = self.info_weight_transformer_.fit_transform(
-                    token_cooccurence
-                )
-            if self.remove_effects_transformer_ is not None:
-                token_cooccurence = self.remove_effects_transformer.fit_transform(
-                    token_cooccurence
-                )
-
-            if i == 0:
-                self.vocabulary_size_ = len(vectorizer_.token_dictionary_)
-                self.representation_ = token_cooccurence
-            else:
-                self.representation_ = hstack([self.representation_, token_cooccurence])
-
-            column_dictionary_ = {
-                (item[0] + i * self.vocabulary_size_): self.vectorizer_names_list_[i]
-                + "_"
-                + item[1]
-                for item in vectorizer_.inverse_token_dictionary_.items()
-            }
-            self.column_dictionary_.update(column_dictionary_)
-
-        self.inverse_column_dictionary_ = {
-            item[1]: item[0] for item in self.column_dictionary_.items()
-        }
-        return self
-
-    def fit_transform(self, X, y=None, **fit_params):
-        self.fit(X)
-        return self.representation_
-
-
-# Takes a sequence of tokens.
-# Takes a list of paramters to be passed to TokenCooccurence iteratively
-# info weight and remove effects
-# hstack them all and return
 
 
 class WordVectorizer(BaseEstimator, TransformerMixin):
@@ -225,7 +84,6 @@ class WordVectorizer(BaseEstimator, TransformerMixin):
         remove_effects_transformer="default",
         remove_effects_transformer_kwds=None,
         normalize=True,
-        ordered_cooccurrence=True,
         dedupe_sentences=True,
     ):
         # make sure we pass the before/after as a switch
@@ -236,13 +94,16 @@ class WordVectorizer(BaseEstimator, TransformerMixin):
         self.token_contractor_kwds = token_contractor_kwds
         self.vectorizer = vectorizer
         self.vectorizer_kwds = vectorizer_kwds
+        #TODO: do I expose this here and update all the
+        # MultiTokenCooccurrenceVectorizer kwds?
+        # Or drop it an let folks pass this in vectorizer_kwds
+        # (JH prefers dropping it and remove_effects)
         self.info_weight_transformer = info_weight_transformer
         self.info_weight_transformer_kwds = info_weight_transformer_kwds
         self.remove_effects_transformer = remove_effects_transformer
         self.remove_effects_transformer_kwds = remove_effects_transformer_kwds
         # Switches
         self.return_normalized = normalize
-        self.ordered_cooccurence = ordered_cooccurrence
         self.dedupe_sentences = dedupe_sentences
 
     def fit(self, X, y=None, **fit_params):
@@ -279,7 +140,9 @@ class WordVectorizer(BaseEstimator, TransformerMixin):
             "contractor",
         )
         if self.token_contractor_ is not None:
-            tokens_by_sentence = self.token_contractor_.fit_transform(tokens_by_sentence)
+            tokens_by_sentence = self.token_contractor_.fit_transform(
+                tokens_by_sentence
+            )
 
         # DEDUPE
         if self.dedupe_sentences:
@@ -297,6 +160,7 @@ class WordVectorizer(BaseEstimator, TransformerMixin):
         # NORMALIZE
         if self.return_normalized:
             self.representation_ = normalize(self.representation_, norm="l1", axis=1)
+
 
         # For ease of finding we promote the token dictionary to be a full class property.
         self.token_dictonary_ = self.vectorizer_.token_dictionary_
